Deprecate old PMM, have new one start allocating at the right spot.
