+++ kernel/include/paging.h
@@ -1,7 +1,7 @@
 #ifndef PAGING_H_
 #define PAGING_H_
 
-#include <pmm_new.h>
+#include <pmm.h>
 
     uint32_t* page_directory;
     uint32_t * first_page_table;
