+++ kernel/include/pmm.h
@@ -1,37 +1,21 @@
-#ifndef PMM_H_
-#define PMM_H_
+#ifndef PMM_NEW_H_
+#define PMM_NEW_H_
 
-//! 8 blocks per byte
-#define BPB 8
-
-//! block size (4k)
-#define BSZ	4096
-
-//! block alignment
-#define BALIGN	BSZ
-
-#define physical_addr uintptr_t
-
-#include <stdint.h>
 #include <stddef.h>
-#include <stdbool.h>
+#include <stdint.h>
+
+#define PAGE_SIZE 4096
+#define ROUND_TO_PAGES(Size) (((uintptr_t)(Size) + PAGE_SIZE - 1) & ~(PAGE_SIZE - 1))
+
+    struct frame {
+        struct frame* next;
+    };
 
 
-    void pmm_init(size_t memSize, uint32_t bitmap);
-    void* pmm_allocate();
-    void pmm_deallocate(void* p);
-    bool pmm_test (int bit);
-    uint32_t memory_size;
-    uint32_t used;
-    uint32_t max;
-    uint32_t* manager_map;
-    void set_bit (int bit);
-    void unset_bit(int bit);
-    int first_free ();
-    uint32_t blocks();
-    void init_region (uint32_t base, size_t size);
-    void deinit_region (uint32_t base, size_t size);
-    void map();
+    struct frame* first_free_addr;
 
-
+    void pmm_init(uintptr_t start_addr,uint32_t length_bytes);
+    void* kmalloc();
+    void kfree(void* address);
+    void pmm_map();
 #endif
