+++ kernel/mem/paging.c
@@ -3,24 +3,23 @@
 
 extern void loadPageDirectory(unsigned int*);
 extern void enablePaging();
-extern void* end;
 void init_paging() {
-    kputs("Init paging");
-    page_directory = &end;
+    /*kputs("Init paging");
+    page_directory = pmm_allocate();
     for(int i=0; i< 1024; i++) {
         page_directory[i] = 0x00000002;
     }
     kputs("Dir ");
-/*    first_page_table = kmalloc();;
+   first_page_table = pmm_allocate();
     for(int j = 0; j< 1024; j++) {
         first_page_table[j] = (j*0x1000) | 3;
     }
     page_directory[0] = ((unsigned int) first_page_table) | 3;
     kputs("Table ");
-    */
-    loadPageDirectory(page_directory);
-//    enablePaging();
-    kputs("K\n");
 
+    loadPageDirectory(page_directory);
+    enablePaging();
+    kputs("K\n");
+*/
 
 }
